fix: renaming svav to svaw
