+++ public/favicons/svaw/apple-touch-icon-light.png
